You can AND the truth table and get intersection, which is AND in the arity=2 case.
But you can also OR the truth table and get something different..but
which is also AND in the arity=2 case! 8>

Imma call it anysection 8>
XD

(I actually needed this! It's for conflict resolution—determining if N
sets are all disjoint, and finding the members that overlap between *any
combination of sets*! in O(N*M) time :D )