Like the javadoc says, I added a new function XD
(Although there was a gaping hole in the previous one in that it didn't
escape NUL characters!!  Which I'm deciding is fine to edit its syntax
to accommodate, because there can't be anything in use that relies on it
not escaping NUL chars or it would've crashed! XD'' )

Also added checks for illegal basenames! (the impetus for fsescaping in
the first place)  :D