***OH MY GOD*** We completely and totally handled ".." path elements wrong in both normpath() and realpath()!!!  X"D
dsjkdflkdj okay now they're right 8))''

But for the not-necessarily-filesystem normpath() there really needs to
be *two* ways of doing it!!

• The normal way which handles the case like a normal posix filesystem
which allows for symlinks, where /a/b/.. might not be /a!  (eg, if /a/b
is a symlink to /c then /a/b/.. is / !)

• And the original way which effectively pretends there's no such thing
as symlinks and thus no system call needs to be made to figure out what
".." is!

This is, for example, how URL's are explicitly specified to work!

https://www.rfc-editor.org/rfc/rfc3986#section-3.3
	However, unlike in a file
	system, these dot-segments are only interpreted within the URI path
	hierarchy and are removed as part of the resolution process (Section
	5.2).