AAAAAAAaaaaaaaand add a general event multicaster system since I referenced the possibility of such, in another project!! 8>
Ironically the creating of this obviates the reason it was created XDDD

It was made so that APIs can use set()/get() not add()/remove()!  But..having made this class, they can use add()/remove() and just expect the *implementations* to use this class! XDD

And I realized while making it that AWT already. does. *exactly that in every way!!* XDD
(Except its implementation is less efficient than the one here :333 )